refactor(data-request): move create dialog out of admin/tickets for shared use
